Fixed NullReferenceExceptions if the database context was null.

+++ src/EntityFramework/DropCreateDatabaseAlways`.cs
@@ -24,8 +24,16 @@ namespace System.Data.Entity
         ///     Executes the strategy to initialize the database for the given context.
         /// </summary>
         /// <param name = "context">The context.</param>
+        /// <exception cref="ArgumentNullException">
+        /// <paramref name="context"/> is <c>null</c>.
+        /// </exception>
         public void InitializeDatabase(TContext context)
         {
+            if (context == null)
+            {
+                throw new ArgumentNullException("context");
+            }
+
             context.Database.Delete();
             context.Database.Create();
             Seed(context);
